backend/src/main/java/br/com/entelgy/model/Candidate.java
package br.com.entelgy.model;

import org.springframework.data.annotation.Id;
import org.springframework.data.mongodb.core.mapping.Document;

/**
 * Candidate model.
 * 
 * @author falvojr
 */
@Document(collection = "candidates")
public class Candidate {
	@Id
	private String id;
	private String name;
	private String photo;
	private String overview;

	public String getId() {
		return id;
	}

	public void setId(String id) {
		this.id = id;
	}

	public String getName() {
		return name;
	}

	public void setName(String name) {
		this.name = name;
	}

	public String getPhoto() {
		return photo;
	}

	public void setPhoto(String photo) {
		this.photo = photo;
	}

	public String getOverview() {
		return overview;
	}

	public void setOverview(String overview) {
		this.overview = overview;
	}

}
